pyproject.toml
[tool.poetry]
name = "simple-image-uploader-backend"
version = "1.0"
description = "Fun project on robin for uploading and previewing images"
authors = ["Andriy Sydorenko <sidorenkoandrij217@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
robyn = "^0.64.1"
psycopg2-binary = "^2.9.10"
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
pre-commit = "^4.0.1"
cloudinary = "^1.41.0"
python-dotenv = "^1.0.1"
pydantic = {extras = ["email"], version = "^2.10.1"}
bcrypt = "^4.2.1"
pyjwt = "^2.10.0"
psycopg2 = "^2.9.10"

[tool.black]
line-length = 119
target-version = ['py312']
include = '\.py$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | \.alembic
)/
'''

[tool.isort]
profile = "black"
